Fix: grafana-cli should receive flags from the command line
grafana-cli should allow configuration overrides to be received from the command line. e.g.

```
grafana-cli admin reset-password cfg:default.paths.logs=custom/log/directory/
```

Regularly, the use of `flag.Args()` should suffice to obtain the arguments passed to the binary. However, since we're using a CLI library to manage the binary, the arguments get manipulated by it and our code ends up not receiving the supplied arguments. Per [the CLI library parsing code](https://github.com/wercker/codegangsta-cli/blob/master/app.go#L123) we can use the context to obtain the tailing arguments.

Additionally, it'll be useful for the user to know whenever these are being overriden or not - hence the addition of logging the configuration to be used as we run the command.